Extract component <FadeInContentTransition>
